Closed url param flag
